requirements.txt
aiogram==2.20
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
frozenlist==1.3.0
google-auth==2.6.6
google-auth-oauthlib==0.5.1
gspread==4.0.1
gspread-asyncio==1.5.3
idna==3.3
multidict==6.0.2
oauthlib==3.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
SimpleQIWI==0.7
six==1.16.0
urllib3==1.26.9
uuid==1.30
yarl==1.7.2
